BEL-3936 Ensure that ECS scaling will not kill running work
In order to ensure our ECS scaling does not kill running jobs, we need to
set task protection while workers are doing work.

Do this by sharing a mutex across all running threads, keeping a count of runnings jobs
in the perform state. When the count gets to 0, unset protected status, and when it increments past 0,
set protected status.