QTY-11338 First draft of debouncing task protection
We're running up against task protection API limits due to quick bursts
of many small tasks. In order to do this, we want to debounce unprotecting tasks
and make sure we never double-send updates to task protection. We still, however
want to protect a task as soon as we start one.

This is a first draft at doing this. I suspect we can simplify and make this more
comprehensible.