BEL-4804 Deal with second level of N+1s
There were 3-4 levels of n+1 bugs in this section of code. Fixing 1
of them seems to get us to the point of not triggering the 55s timeout,
but we stretch goaled this in.

Canvas forbids using .includes(), presumably because they prefer to
always make the explicit choice of which behaviour is happening to
improve performance, so we have to use preload or eager_load here.

We're choosing preload, because everything that is using this is close by,
and that seems less confusing.

Co-authored-by: Leonid Medovyy <leonid.medovyy@strongmind.com>